tests/test_models/test_place.py
#!/usr/bin/python3
"""
Unittest for Place
"""
import unittest
from models.place import Place


class TestPlace(unittest.TestCase):
    """Defines test cases for the Place class."""
    
    def setUp(self):
        """Setup for Place tests."""
        self.place_1 = Place()
        
    def test_instance(self):
        """Test if the object is an instance of Place."""
        self.assertIsInstance(self.place_1, Place)
        
    def test_attributes(self):
        """Test Place attributes."""
        # Add assertions for all unique Place attributes
        self.assertTrue(hasattr(self.place_1, "city_id"))
        self.assertTrue(hasattr(self.place_1, "user_id"))
        # Continue for all attributes
        
if __name__ == '__main__':
    unittest.main()
